tests/wpt/tests/editing/other/insert-list-preserving-selection.tentative.html

<!doctype html>
<html>
<head>
<meta chareset="utf-8">
<meta name="timeout" content="long">
<meta name="variant" content="?styleWithCSS=false&command=insertOrderedList">
<meta name="variant" content="?styleWithCSS=false&command=insertUnorderedList">
<meta name="variant" content="?styleWithCSS=true&command=insertOrderedList">
<meta name="variant" content="?styleWithCSS=true&command=insertUnorderedList">
<title>Test preserving selection after insert*List</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="../include/editor-test-utils.js"></script>
</head>
<body>
<div contenteditable></div>
<script>
"use strict";

const editor = document.querySelector("div[contenteditable]");
const utils = new EditorTestUtils(editor);
const searchParams = new URLSearchParams(document.location.search);
const styleWithCSS = searchParams.get("styleWithCSS");
const command = searchParams.get("command");
document.execCommand("styleWithCSS", false, styleWithCSS);

// Note that it's not scope of this test how browsers to convert the selected
// content to a list.

// html: Initial HTML which will be set editor.innerHTML, it should contain
//       selection range with a pair of "[" or "{" and "]" or "}".
// expectedSelectedString: After executing "outdent", compared with
//                         getSelection().toString().replace(/[ \n\r\t]+/g, "")
const tests = [
  {
    html: "<div>a[b]c</div>",
    expectedSelectedString: "b",
  },
  {
    html: "<div>a[bc</div><div>de]f</div>",
    expectedSelectedString: "bcde",
  },
  {
    html: "<div>a[bc<br>de]f</div>",
    expectedSelectedString: "bcde",
  },
  {
    html: "<ul><li>a[b]c</li></ul>",
    expectedSelectedString: "b",
  },
  {
    html: "<ul><li>a[bc</li><li>de]f</li></ul>",
    expectedSelectedString: "bcde",
  },
  {
    html: "<ul><li>a[bc</li><li>de]f</li><li>ghi</li></ul>",
    expectedSelectedString: "bcde",
  },
  {
    html: "<ul><li>abc</li><li>d[ef</li><li>gh]i</li></ul>",
    expectedSelectedString: "efgh",
  },
  {
    html: "<ol><li>a[b]c</li></ol>",
    expectedSelectedString: "b",
  },
  {
    html: "<ol><li>a[bc</li><li>de]f</li></ol>",
    expectedSelectedString: "bcde",
  },
  {
    html: "<ol><li>a[bc</li><li>de]f</li><li>ghi</li></ol>",
    expectedSelectedString: "bcde",
  },
  {
    html: "<ol><li>abc</li><li>d[ef</li><li>gh]i</li></ol>",
    expectedSelectedString: "efgh",
  },
  {
    html: "<ul><li>a[bc</li></ul>" +
          "<ul><li>de]f</li></ul>",
    expectedSelectedString: "bcde",
  },
  {
    html: "<ol><li>a[bc</li></ol>" +
          "<ul><li>de]f</li></ul>",
    expectedSelectedString: "bcde",
  },
  {
    html: "<ul><li>a[bc</li></ul>" +
          "<ol><li>de]f</li></ol>",
    expectedSelectedString: "bcde",
  },
  {
    html: "<div>a[bc</div>" +
          "<ul><li>de]f</li><li>ghi</li></ul>",
    expectedSelectedString: "bcde",
  },
  {
    html: "<ul><li>abc</li><li>d[ef</li></ul>" +
          "<div>gh]i</div>",
    expectedSelectedString: "efgh",
  },
  {
    html: "<div>a[bc</div>" +
          "<ol><li>de]f</li><li>ghi</li></ol>",
    expectedSelectedString: "bcde",
  },
  {
    html: "<ol><li>abc</li><li>d[ef</li></ol>" +
          "<div>gh]i</div>",
    expectedSelectedString: "efgh",
  },
  {
    html: "<table><tr><td>a[b]c</td></tr></table>",
    expectedSelectedString: "b",
  },
  {
    html: "<table><tr><td>a[bc</td><td>de]f</td></tr></table>",
    expectedSelectedString: "bcde",
  },
  {
    html: "<table><tr><td>a[bc</td></tr><tr><td>de]f</td></tr></table>",
    expectedSelectedString: "bcde",
  },
  {
    html: "<div>a[bc</div>" +
          "<table><tr><td>de]f</td></tr></table>",
    expectedSelectedString: "bcde",
  },
  {
    html: "<table><tr><td>a[bc</td></tr></table>" +
          "<div>de]f</div>",
    expectedSelectedString: "bcde",
  },
];

for (const t of tests) {
  test(() => {
    utils.setupEditingHost(t.html);
    document.execCommand(command);
    assert_equals(
      getSelection().toString().replace(/[ \n\r\t]+/g, ""),
      t.expectedSelectedString,
      `Result: ${editor.innerHTML}`
    );
  }, `Preserve selection after ${command} at ${t.html}`);
}

</script>
</body>
</html>
